Update rules.yaml
on OCP clusters where the cluster monitoring operator is used (as advised in README) instead of the user-workload-monitoring stack, the `namespace` in the metrics is always the one of the cert-utils-operator. The namespace of the secret/certificate gets rewritten to exported_namespace by default.
Not 100% sure if this is the correct place to fix this though?